fix: remove empty REMOVE operations when diffing from undefined
When diffing from undefined to a value, the diff function was generating an
empty REMOVE operation that didn't have a value property. This was unnecessary
since there's nothing to remove when starting from undefined.

Modified the condition in the compare function to only add a REMOVE operation
if the old object is not undefined.

Fixes #204